app/backend/api/requirements.txt
flask
flask_sqlalchemy
flask_cors
flask_mail
flask_bcrypt
flask_jwt_extended
python-dotenv
